pyproject.toml
[project]
name = "tcss456"
version = "0.1.0"
description = "dependencies needed for Sassy P.A"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "llama-cpp-python>=0.3.8",
    "python-dotenv>=1.1.0",
    "transformers>=4.51.3",
    "torch>=2.6.0",
    "pydantic>=2.10",
    "uvicorn>=0.34.3",
    "asyncio>=3.4.3",
]
